Updated JSON file and word validity
There were no 1-letter words in the JSON file so I recompiled it to eliminate those entries.

Also added a function to check if a word is too short or too long and raises an error if it is either of those.

Also updated the README to reflect the changes made to how the program works.